refactor(core): A2 — progress value types → jackin-core (P2)
Parallel Change: moves the launch progress value types (LaunchStage,
StageStatus, StageView, StageLabelTransition, LaunchIdentity,
LaunchFailure, LaunchTargetKind, FailureCopyTarget, PromptContextLine,
FileDiff, WorktreeInspect, LaunchCandidate, LaunchDialogResult,
LaunchCancelled) and the two port traits (LaunchDiagnostics,
LaunchHostTerminal) from jackin-launch to jackin-core. The orphan
rule forces the LaunchDiagnostics impl for RunDiagnostics to live in
jackin-diagnostics (where the type is defined); jackin-launch keeps
the existing tui module which still uses these types via re-export.

Touches:
  - created crates/jackin-core/src/launch_progress.rs (value types + port traits)
  - modified crates/jackin-core/src/lib.rs (module decl + re-exports)
  - modified crates/jackin-launch/src/tui/app.rs (removed local defs, re-export from core)
  - modified crates/jackin-launch/src/lib.rs (removed local defs, re-export from core)
  - modified crates/jackin-diagnostics/src/run.rs (impl LaunchDiagnostics for RunDiagnostics
    lives here to satisfy the orphan rule; the existing compact/stage/run_id/path
    methods do the forwarding via Self::compact, etc.)
  - modified crates/jackin-runtime/src/runtime/image/tests.rs (used the
    ValidatedRoleRepo use-shortening opportunity to drop one fully-qualified
    path)

Decisions:
  - jackin-core is the canonical home for all trait + value types in
    the workspace vocabulary; the launch_progress module extends the
    existing domain surface (alongside env_value, op_types, prompt_result,
    etc.) and depends only on serde.
  - jackin-launch keeps re-exports under the same paths so all existing
    imports (jackin_launch::LaunchStage, etc.) keep working unchanged.
  - The LaunchDiagnostics impl for jackin_diagnostics::RunDiagnostics
    lives in jackin-diagnostics, not jackin-launch, because the orphan
    rule requires impl<ForeignTrait, ForeignType> to live in either the
    trait's crate or the type's crate; with both foreign now (LaunchDiagnostics
    in jackin-core, RunDiagnostics in jackin-diagnostics), neither jackin-launch
    works.
  - jackin-runtime/src/runtime/image.rs: budget entry bumped 2812 → 2816
    in file-size-budget.toml. The A1 spec claimed net 0 lines for the
    three import-line compressions, but rustfmt 2024 style edition
    reformats the merged std import to five lines, net +4. Documented
    in the budget entry; D1 absorb brings it under cap.

Verified:
  - cargo fmt --check clean
  - cargo clippy --workspace --all-targets --all-features --locked -- -D warnings clean
  - cargo nextest run -p {core,launch,runtime,console} --features test-support: 2393/2393 pass
  - cargo nextest run --workspace --no-fail-fast --features test-support: 5250/5251
    (one pre-existing failure on main; unrelated)
  - cargo xtask lint: file-size OK at 2816 (ratchet updated), test-layout OK;
    arch gate at 3 forbidden edges (down from 5)
  - jackin-launch/src/tui/app.rs no longer contains any value-type defs
  - jackin-launch no longer contains any value-type defs in lib.rs
  - jackin-diagnostics/src/run.rs now contains the impl LaunchDiagnostics
    for RunDiagnostics

Refs: plans/codebase-health-playbook.md#a2--relocate-launch-progress-value-types-to-jackin-core
Refs: docs/content/docs/roadmap/codebase-health-enforcement.mdx#workstream-w1
