+++ crates/jackin-core/src/lib.rs
@@ -20,6 +20,7 @@ pub mod env_value;
 pub mod instance;
 pub mod isolation;
 pub mod isolation_record;
+pub mod launch_progress;
 pub mod manifest;
 pub mod op_cache;
 pub mod op_reference;
@@ -49,6 +50,12 @@ pub use docker_security::{
 pub use env_value::{EnvValue, Extended, FieldTarget, OpRef};
 pub use isolation::{MountIsolation, ParseMountIsolationError};
 pub use isolation_record::{CleanupStatus, DriftDetection, IsolationRecord};
+pub use launch_progress::{
+    FailureCopyTarget, FileDiff, LaunchCancelled, LaunchCandidate, LaunchDiagnostics,
+    LaunchDialogResult, LaunchFailure, LaunchHostTerminal, LaunchIdentity, LaunchStage,
+    LaunchTargetKind, PromptContextLine, StageLabelTransition, StageStatus, StageView,
+    WorktreeInspect,
+};
 pub use path_text::shorten_home;
 pub use paths::JackinPaths;
 pub use prompt_result::PromptResult;
