+++ docs/content/docs/roadmap/codebase-health-enforcement.mdx
@@ -225,7 +225,7 @@ Used by every C/E slice. Pure move; contents are not rewritten.
 
 - [x] **A0 — `PromptResult` → `jackin-core`.** *(shipped)* Broke the `jackin-env → jackin-launch` edge for this type.
 - [x] **A1 — `build_log` trait port (D2).** Moves: define `pub trait BuildLogSink` in `jackin-core`; `jackin-docker` shell runner takes `&dyn BuildLogSink` and calls `push_line`; `jackin-launch(-tui)` implements it; `jackin-runtime` injects the UI sink at the call site. How: [Branch by Abstraction](https://martinfowler.com/bliki/BranchByAbstraction.html). Verify: build-log pane renders identically; tests green. Done when `jackin-docker` no longer names `jackin_launch`.
-- [ ] **A2 — `progress` value types → `jackin-core`.** Moves: relocate the launch `progress` value types out of `jackin-launch`; the streaming progress updates use the same port shape as A1. Done when `jackin-runtime`/`jackin-env` no longer depend on `jackin-launch` for types.
+- [x] **A2 — `progress` value types → `jackin-core`.** Moves: relocate the launch `progress` value types out of `jackin-launch`; the streaming progress updates use the same port shape as A1. Done when `jackin-runtime`/`jackin-env` no longer depend on `jackin-launch` for types.
 - [ ] **A3 — presentation helpers out of `jackin-core` (P7).** Moves: `ansi_text`, `prune_output`, `url_text` → `jackin-tui` (presentation). Done when `jackin-core` has no rendering code.
 - [ ] **A4 — terminal-ownership state out of `jackin-diagnostics` (P7).** Moves: `rich_surface_active` / `host_screen_owned` state → a presentation/terminal crate; `jackin-diagnostics` keeps observability only. Done when no domain crate depends on `jackin-diagnostics` for terminal state.
 - [ ] **A5 — land the W4 dependency-direction bans + invariant headers.** Prereq: A1–A4. Adds the `cargo-deny` bans (`jackin-docker`/`jackin-env`/`jackin-runtime` ↛ presentation; domain ↛ infra/presentation) and the per-crate `//!` headers. This freezes the boundary fixes so they cannot regress; every later phase runs against it.
